# Conflicts:
#	SlipeServer.Server.TestTools/TestingClient.cs
#	SlipeServer.Server.Tests/Integration/HostingTests.cs

+++ SlipeServer.Server.TestTools/TestingClient.cs
@@ -12,12 +12,12 @@ namespace SlipeServer.Server.TestTools;
 
 public class TestingClient : Client<Player>
 {
-    public uint Address { get; private set; }
+    public ulong Address { get; private set; }
 
     public event Action<TestingClient, Packet> PacketSent;
-
-    public TestingClient(uint address, INetWrapper netWrapper, Player player)
+    public TestingClient(ulong address, INetWrapper netWrapper, Player player)
         : base(address, netWrapper, player)
+
     {
         this.Address = address;
         this.ConnectionState = Enums.ClientConnectionState.Joined;
